Revert "added citation_?helper and user_mailer to stash datacite"
This reverts commit e2a84ef0fd509ae176d98fefddd403dd16e6188d.
