+++ stash_datacite/app/views/stash_datacite/resources/_review.html.erb
@@ -11,7 +11,11 @@
     </div>
 
     <section class="o-section">
-      <%= render partial: "stash_datacite/shared/citations", locals: { resource: @resource } %>
+    <%= render partial: "stash_datacite/shared/citations", locals: { authors: @review.authors, title: @review.title_str,
+                         resource_type: @review.resource_type, version: "v#{@resource.version_number}",
+                         identifier: @resource.identifier.nil? ? nil : "#{@review.identifier.identifier}",
+                         publisher: @review.publisher,
+                         publication_years: @resource.publication_years } %>
     </section>
 
     <%= render partial: "stash_datacite/descriptions/show", locals:
