+++ stash_datacite/app/views/stash_datacite/resources/_show.html.erb
@@ -8,7 +8,13 @@
   </div>
 
   <section class="o-section">
-  <%= render partial: "stash_datacite/shared/citations", locals: { resource: resource } %>
+  <%= render partial: "stash_datacite/shared/citations",
+             locals: { authors: review.authors, title: review.title_str,
+                       resource_type: review.resource_type,
+                       version: resource.stash_version.nil? ? 'v0' : "v#{review.version.version }",
+                       identifier: resource.identifier.nil? ? 'DOI' : "#{review.identifier.identifier }",
+                       publisher: review.publisher,
+                       publication_years: resource.publication_years } %>
   </section>
   <%= render partial: "stash_datacite/descriptions/show", locals: { abstract: review.abstract,
                                                                     methods: review.methods, other: review.other } %>
