+++ stash_datacite/app/views/stash_datacite/shared/_citations.html.erb
@@ -1,2 +1,2 @@
 <h3 class="o-heading__level2">Citation</h3>
-<p><%= citation(resource) %></p>
+<p><%= citation(authors, title, resource_type, version, identifier, publisher, publication_years) %></p>
