+++ stash_engine/app/views/stash_engine/user_mailer/published.html.erb
